created presentation domain data and used dumy data to get basic android app to work

+++ src/commonMain/kotlin/domain/weather/entities/Weather.kt
@@ -2,5 +2,5 @@ package domain.weather.entities
 
 data class Weather (
     val status: String? = null,
-    val temperature: Int? = null
+    val temperature: Double? = null
 )
